+++ secret/src/lib.rs
@@ -1,26 +1,33 @@
-//! # Secret
-//!
-//! The core concept of this library is to abstract the concept of
-//! secret. A secret can be retrieved either from a raw string, from a
-//! command or from a keyring entry. The associated structure is
-//! [`Secret`].
+#![doc = include_str!("../README.md")]
 
 #[cfg(feature = "derive")]
-pub mod derive;
+pub(crate) mod derive;
 mod error;
 
 #[cfg(feature = "keyring")]
 pub use keyring;
 #[cfg(feature = "keyring")]
 use keyring::KeyringEntry;
-use log::debug;
 #[cfg(feature = "command")]
 pub use process;
 #[cfg(feature = "command")]
 use process::Command;
+use tracing::debug;
 
 #[doc(inline)]
 pub use crate::error::{Error, Result};
+
+#[cfg(any(
+    all(feature = "tokio", feature = "async-std"),
+    not(any(feature = "tokio", feature = "async-std"))
+))]
+compile_error!("Either feature \"tokio\" or \"async-std\" must be enabled for this crate.");
+
+#[cfg(any(
+    all(feature = "rustls", feature = "openssl"),
+    not(any(feature = "rustls", feature = "openssl"))
+))]
+compile_error!("Either feature \"rustls\" or \"openssl\" must be enabled for this crate.");
 
 /// The secret.
 ///
@@ -30,33 +37,40 @@ pub use crate::error::{Error, Result};
 #[cfg_attr(
     feature = "derive",
     derive(serde::Serialize, serde::Deserialize),
-    serde(rename_all = "kebab-case"),
-    serde(from = "derive::Secret")
+    serde(from = "derive::Secret"),
+    serde(rename_all = "kebab-case")
 )]
 pub enum Secret {
-    /// The secret is contained in a raw string, usually not safe to
-    /// use and so not recommended.
+    /// The secret is empty.
+    #[default]
+    Empty,
+
+    /// The secret is contained in a raw string.
+    ///
+    /// This variant is not safe to use and therefore not
+    /// recommended. Yet it works well for testing purpose.
     Raw(String),
 
     /// The secret is exposed by the given shell command.
+    ///
+    /// This variant takes the secret from the first line returned by
+    /// the given shell command.
+    ///
+    /// See [process-lib](https://crates.io/crates/process-lib).
     #[cfg(feature = "command")]
     #[cfg_attr(feature = "derive", serde(alias = "cmd"))]
     Command(Command),
 
-    /// The secret is contained in the given user's global keyring at
-    /// the given entry.
+    /// The secret is contained in the user's global keyring at the
+    /// given entry.
+    ///
+    /// See [keyring-lib](https://crates.io/crates/keyring-lib).
     #[cfg(feature = "keyring")]
-    #[cfg_attr(feature = "derive", serde(rename = "keyring"))]
-    KeyringEntry(KeyringEntry),
-
-    /// The secret is not defined.
-    #[default]
-    #[cfg_attr(feature = "derive", serde(skip_serializing))]
-    Undefined,
+    Keyring(KeyringEntry),
 }
 
 impl Secret {
-    /// Create a new undefined secret.
+    /// Create a new empty secret.
     pub fn new() -> Self {
         Default::default()
     }
@@ -75,7 +89,7 @@ impl Secret {
     /// Create a new secret from the given keyring entry.
     #[cfg(feature = "keyring")]
     pub fn new_keyring_entry(entry: KeyringEntry) -> Self {
-        Self::KeyringEntry(entry)
+        Self::Keyring(entry)
     }
 
     /// Try to create a new secret from the given entry.
@@ -83,13 +97,13 @@ impl Secret {
     pub fn try_new_keyring_entry(
         entry: impl TryInto<KeyringEntry, Error = keyring::Error>,
     ) -> Result<Self> {
-        let entry = entry.try_into().map_err(Error::KeyringError)?;
-        Ok(Self::KeyringEntry(entry))
+        let entry = entry.try_into()?;
+        Ok(Self::new_keyring_entry(entry))
     }
 
-    /// Return `true` if the secret is not defined.
-    pub fn is_undefined(&self) -> bool {
-        matches!(self, Self::Undefined)
+    /// Return `true` if the secret is empty.
+    pub fn is_empty(&self) -> bool {
+        *self == Self::Empty
     }
 
     /// Get the secret value.
@@ -99,48 +113,66 @@ impl Secret {
     /// from the global keyring using its inner key.
     pub async fn get(&self) -> Result<String> {
         match self {
-            Self::Raw(raw) => Ok(raw.clone()),
-            #[cfg(feature = "command")]
-            Self::Command(cmd) => Ok(cmd
-                .run()
-                .await
-                .map_err(Error::GetSecretFromCommand)?
-                .to_string_lossy()
-                .lines()
-                .take(1)
-                .next()
-                .ok_or(Error::GetSecretFromCommandEmptyOutputError)?
-                .to_owned()),
-            #[cfg(feature = "keyring")]
-            Self::KeyringEntry(entry) => {
-                Ok(entry.get_secret().await.map_err(Error::KeyringError)?)
+            Self::Empty => {
+                return Err(Error::GetEmptySecretError);
             }
-            Self::Undefined => Err(Error::GetUndefinedSecretError),
+            Self::Raw(secret) => {
+                return Ok(secret.clone());
+            }
+            #[cfg(feature = "command")]
+            Self::Command(cmd) => {
+                let full_secret = cmd
+                    .run()
+                    .await
+                    .map_err(Error::GetSecretFromCommand)?
+                    .to_string_lossy();
+
+                let first_line_secret = full_secret
+                    .lines()
+                    .take(1)
+                    .next()
+                    .ok_or(Error::GetSecretFromCommandEmptyOutputError)?
+                    .to_owned();
+
+                Ok(first_line_secret)
+            }
+            #[cfg(feature = "keyring")]
+            Self::Keyring(entry) => {
+                let secret = entry.get_secret().await?;
+                Ok(secret)
+            }
         }
     }
 
     /// Find the secret value.
     ///
     /// Like [`get`], but returns [`None`] if the secret value is not
-    /// found or undefined.
+    /// found or empty.
     pub async fn find(&self) -> Result<Option<String>> {
         match self {
-            Self::Raw(secret) => Ok(Some(secret.clone())),
-            #[cfg(feature = "command")]
-            Self::Command(cmd) => Ok(cmd
-                .run()
-                .await
-                .map_err(Error::GetSecretFromCommand)?
-                .to_string_lossy()
-                .lines()
-                .take(1)
-                .next()
-                .map(ToOwned::to_owned)),
-            #[cfg(feature = "keyring")]
-            Self::KeyringEntry(entry) => {
-                Ok(entry.find_secret().await.map_err(Error::KeyringError)?)
+            Self::Empty => {
+                return Ok(None);
             }
-            Self::Undefined => Ok(None),
+            Self::Raw(secret) => {
+                return Ok(Some(secret.clone()));
+            }
+            #[cfg(feature = "command")]
+            Self::Command(cmd) => {
+                let full_secret = cmd
+                    .run()
+                    .await
+                    .map_err(Error::GetSecretFromCommand)?
+                    .to_string_lossy();
+
+                let first_line_secret = full_secret.lines().take(1).next().map(ToOwned::to_owned);
+
+                Ok(first_line_secret)
+            }
+            #[cfg(feature = "keyring")]
+            Self::Keyring(entry) => {
+                let secret = entry.find_secret().await?;
+                Ok(secret)
+            }
         }
     }
 
@@ -149,72 +181,47 @@ impl Secret {
     /// This is only applicable for raw secrets and keyring-based
     /// secrets. A secret value cannot be changed for command-base
     /// secrets, since the value is the output of the command.
-    pub async fn set(&mut self, secret: impl AsRef<str>) -> Result<String> {
-        let secret = secret.as_ref();
-
+    pub async fn set(&mut self, secret: impl ToString) -> Result<String> {
         match self {
             Self::Raw(prev) => {
-                *prev = secret.to_owned();
+                *prev = secret.to_string();
             }
             #[cfg(feature = "command")]
             Self::Command(_) => {
                 debug!("cannot change value of command-based secret");
             }
             #[cfg(feature = "keyring")]
-            Self::KeyringEntry(entry) => {
-                entry
-                    .set_secret(secret)
-                    .await
-                    .map_err(Error::KeyringError)?;
-            }
-            Self::Undefined => {
-                debug!("cannot change value of undefined secret");
+            Self::Keyring(entry) => entry.set_secret(secret.to_string()).await?,
+            Self::Empty => {
+                debug!("cannot change value of empty secret");
             }
         }
 
-        Ok(secret.to_owned())
+        Ok(secret.to_string())
     }
 
     /// Change the secret value of the keyring-based secret only.
     ///
     /// This function as no effect on other secret variants.
     #[cfg(feature = "keyring")]
-    pub async fn set_only_keyring(&self, secret: impl AsRef<str>) -> Result<String> {
-        let secret = secret.as_ref();
-
-        if let Self::KeyringEntry(entry) = self {
-            entry
-                .set_secret(secret)
-                .await
-                .map_err(Error::KeyringError)?;
+    pub async fn set_if_keyring(&self, secret: impl ToString) -> Result<String> {
+        if let Self::Keyring(entry) = self {
+            let secret = secret.to_string();
+            entry.set_secret(&secret).await?;
+            return Ok(secret);
         }
 
-        Ok(secret.to_owned())
+        Ok(secret.to_string())
     }
 
-    /// Replace undefined secret by a keyring-based one.
-    ///
-    /// This function has no effect on other variants.
-    #[cfg(feature = "keyring")]
-    pub fn replace_undefined_to_keyring(
-        &mut self,
-        entry: impl TryInto<KeyringEntry, Error = keyring::Error>,
-    ) -> Result<()> {
-        if self.is_undefined() {
-            *self = Self::try_new_keyring_entry(entry)?
-        }
-
-        Ok(())
-    }
-
-    /// Delete the secret value and make the current secret undefined.
+    /// Delete the secret value and make the current secret empty.
     pub async fn delete(&mut self) -> Result<()> {
         #[cfg(feature = "keyring")]
-        if let Self::KeyringEntry(entry) = self {
-            entry.delete_secret().await.map_err(Error::KeyringError)?;
+        if let Self::Keyring(entry) = self {
+            entry.delete_secret().await?;
         }
 
-        *self = Self::Undefined;
+        *self = Self::Empty;
 
         Ok(())
     }
@@ -223,11 +230,20 @@ impl Secret {
     ///
     /// This function has no effect on other variants.
     #[cfg(feature = "keyring")]
-    pub async fn delete_only_keyring(&self) -> Result<()> {
-        if let Self::KeyringEntry(entry) = self {
-            entry.delete_secret().await.map_err(Error::KeyringError)?;
+    pub async fn delete_if_keyring(&self) -> Result<()> {
+        if let Self::Keyring(entry) = self {
+            entry.delete_secret().await?;
         }
 
         Ok(())
+    }
+
+    /// Replace empty secret variant with the given one.
+    ///
+    /// This function has no effect on other variants.
+    pub fn replace_if_empty(&mut self, new: Self) {
+        if self.is_empty() {
+            *self = new
+        }
     }
 }